Implemented skeleton functionality for RecordedCommunitySet and RecordedCommunitySummarizer, partially integrated in world. AEcoWorld in bloated state as a result.

+++ include/chemical-ecology/Config.hpp
@@ -10,7 +10,7 @@ namespace chemical_ecology {
     GROUP(GLOBAL_SETTINGS, "Global settings"),
     VALUE(SEED, int, -1, "Seed for a simulation"),
     VALUE(N_TYPES, size_t, 9, "Number of types"),
-    VALUE(UPDATES, int, 1000, "Number of time steps to run for"),
+    VALUE(UPDATES, size_t, 1000, "Number of time steps to run for"),
     VALUE(MAX_POP, int, 10000, "Maximum population size for one type in one cell"),
     VALUE(DIFFUSION, double, 1, "Proportion of each population that diffuses to adjacent cells"),
     VALUE(SEEDING_PROB, double, .001, "Probability that a member of a given species will randomly be introduced to a cell"),
